--- src/main/java/kitchenpos/dao/ProductDao.java
@@ -1,14 +1,0 @@
-package kitchenpos.dao;
-
-import kitchenpos.domain.Product;
-
-import java.util.List;
-import java.util.Optional;
-
-public interface ProductDao {
-    Product save(Product entity);
-
-    Optional<Product> findById(Long id);
-
-    List<Product> findAll();
-}
