--- src/main/java/kitchenpos/domain/MenuProduct.java
@@ -1,49 +1,0 @@
-package kitchenpos.domain;
-
-public class MenuProduct {
-    private Long seq;
-    private Long menuId;
-    private Long productId;
-    private long quantity;
-
-    public  MenuProduct() {
-
-    }
-    public MenuProduct(Long menuId, Long productId, long quantity) {
-        this.menuId = menuId;
-        this.productId = productId;
-        this.quantity = quantity;
-    }
-
-    public Long getSeq() {
-        return seq;
-    }
-
-    public void setSeq(final Long seq) {
-        this.seq = seq;
-    }
-
-    public Long getMenuId() {
-        return menuId;
-    }
-
-    public void setMenuId(final Long menuId) {
-        this.menuId = menuId;
-    }
-
-    public Long getProductId() {
-        return productId;
-    }
-
-    public void setProductId(final Long productId) {
-        this.productId = productId;
-    }
-
-    public long getQuantity() {
-        return quantity;
-    }
-
-    public void setQuantity(final long quantity) {
-        this.quantity = quantity;
-    }
-}
