+++ src/main/java/kitchenpos/order/domain/OrderTables.java
@@ -16,7 +16,7 @@ import java.util.stream.Collectors;
 public class OrderTables {
     @OneToMany(cascade = {CascadeType.PERSIST, CascadeType.MERGE}, orphanRemoval = true)
     @JoinColumn(name = "table_group_id")
-    private List<OrderTable> orderTables = new ArrayList<>();
+    private final List<OrderTable> orderTables = new ArrayList<>();
 
     public void add(Long tableGroupId, List<OrderTable> orderTables) {
         if (CollectionUtils.isEmpty(orderTables) || orderTables.size() < 2) {
